Moved simple-arpeggiator-touch.cpp to common dir
